動かなくなったので切り戻し

+++ lib/model/cart.php
@@ -12,10 +12,12 @@
  * @return boolean
  */
 function cart_is_exists_item($db, $user_id, $item_id) {
-	$sql = 'SELECT item_id, amount FROM carts
-	WHERE user_id = ? AND item_id = ?;';
-	$params = array($user_id, $item_id);
-	$cart = db_select($db, $sql, $params);
+	$sql = <<<EOD
+SELECT item_id, amount FROM carts
+ WHERE user_id = {$user_id} AND item_id = {$item_id}
+EOD;
+
+	$cart = db_select($db, $sql);
 	return empty($cart) === false;
 }
 
@@ -25,13 +27,13 @@ function cart_is_exists_item($db, $user_id, $item_id) {
  * @return int | NULL
  */
 function cart_total_price($db, $user_id) {
-	$sql =
-	'SELECT sum(price * amount) as total_price
- 	FROM carts JOIN items
- 	ON carts.item_id = items.id
-	WHERE items.status = 1 AND user_id = ?;';
-	$params = array($user_id); 
-	$row = db_select_one($db, $sql, $params);
+	$sql = <<<EOD
+SELECT sum(price * amount) as total_price
+ FROM carts JOIN items
+ ON carts.item_id = items.id
+ WHERE items.status = 1 AND user_id = {$user_id}
+EOD;
+	$row = db_select_one($db, $sql);
 	if (empty($row)) {
 		return null;
 	}
@@ -44,13 +46,13 @@ function cart_total_price($db, $user_id) {
  * @return array
  */
 function cart_list($db, $user_id) {
-	$sql =
-	'SELECT carts.id, item_id, name, price, img, amount, (price * amount) as amount_price
- 	FROM carts JOIN items
- 	ON carts.item_id = items.id
-	WHERE items.status = 1 AND user_id = ?;';
-	$params = array($user_id); 
-	return db_select($db, $sql, $params);
+	$sql = <<<EOD
+ SELECT carts.id, item_id, name, price, img, amount, (price * amount) as amount_price
+ FROM carts JOIN items
+ ON carts.item_id = items.id
+ WHERE items.status = 1 AND user_id = {$user_id}
+EOD;
+	return db_select($db, $sql);
 }
 
 /**
@@ -63,17 +65,18 @@ function cart_regist($db, $user_id, $item_id) {
 	$sql = '';
 
 	if (cart_is_exists_item($db, $user_id, $item_id)) {
-		$sql =
-		'UPDATE carts
- 		SET amount = amount + 1 , update_date = NOW()
- 		WHERE user_id = ? AND item_id = ?;';
+		$sql = <<<EOD
+UPDATE carts
+ SET amount = amount + 1 , update_date = NOW()
+ WHERE user_id = {$user_id} AND item_id = {$item_id}
+EOD;
 	} else {
-		$sql =
-		'INSERT INTO carts (user_id, item_id, amount, create_date, update_date)
-		VALUES (?, ?, 1, NOW(), NOW());';
+		$sql = <<<EOD
+INSERT INTO carts (user_id, item_id, amount, create_date, update_date)
+VALUES ({$user_id}, {$item_id}, 1, NOW(), NOW())
+EOD;
 	}
-	$params = array($user_id, $item_id);
-	return db_update($db, $sql, $params);
+	return db_update($db, $sql);
 }
 
 /**
@@ -84,12 +87,12 @@ function cart_regist($db, $user_id, $item_id) {
  * @return int
  */
 function cart_update($db, $id, $user_id, $amount) {
-	$sql =
-	'UPDATE carts
- 	SET amount = ?, update_date = NOW()
- 	WHERE id = ? AND user_id = ?;';
-	$params = array($amount, $id, $user_id);
-	return db_update($db, $sql, $params);
+	$sql = <<<EOD
+UPDATE carts
+ SET amount = {$amount}, update_date = NOW()
+ WHERE id = {$id} AND user_id = {$user_id}
+EOD;
+	return db_update($db, $sql);
 }
 
 /**
@@ -99,11 +102,11 @@ function cart_update($db, $id, $user_id, $amount) {
  * @return int
  */
 function cart_delete($db, $id, $user_id) {
-	$sql =
-	'DELETE FROM carts
-	WHERE id = ? AND user_id = ?;';
-	$params = array($id, $user_id); 
-	return db_update($db, $sql, $params);
+	$sql = <<<EOD
+DELETE FROM carts
+ WHERE id = {$id} AND user_id = {$user_id}
+EOD;
+	return db_update($db, $sql);
 }
 
 /**
@@ -112,9 +115,6 @@ function cart_delete($db, $id, $user_id) {
  * @return int
  */
 function cart_clear($db, $user_id) {
-	$sql =
-	'DELETE FROM carts
-	WHERE user_id = ?;';
-	$params = array($user_id);
-	return db_update($db, $sql, $params);
+	$sql = 'DELETE FROM carts WHERE user_id = ' . $user_id;
+	return db_update($db, $sql);
 }
